pyproject.toml
[project]
name = "workflow-doc-agent"
version = "0.1.0"
description = "Auto-document a folder of production work (Python + Excel) using a swappable LLM provider (Gemini or Claude). Cost-aware, eval-checked, ships in one command."
authors = [{ name = "Nader Mohamed" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "google-genai>=0.3.0",
    "openpyxl>=3.1.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
]

[project.scripts]
workflow-doc = "workflow_doc_agent.cli:app"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
